+++ NEWS.md
@@ -9,8 +9,7 @@ Version 27.0-dev (main)
 Version 26.4-dev (v26-fixes)
 ----------------------------
 
-* Increased maximum screen size for using layout for mobile devices
-  from 7 to 10 inches.
+*
 
 Version 26.3 (06 Feb 2025)
 --------------------------