+++ NEWS.md
@@ -4,7 +4,8 @@ Pentobi Release Notes
 Version 26.4-dev (v26-fixes)
 ----------------------------
 
-*
+* Workaround for QTBUG-139679 which broke entering text in comment
+  area in Qt 6.9.2.
 
 Version 26.3 (06 Feb 2025)
 --------------------------